Proposal: stick to one format for date-time time-offset
[https://www.ietf.org/rfc/rfc3339.txt|RFC 3339] allows two formats for
the time-offset of a UTC date-time:
- `Z`
- `+00:00`
I could not find a functional difference between the two.

However, it would cause less confusion down the road if we stick to one
of these formats. PHP's `datetime` extension using `DATE_ATOM` as a
format produces `+00:00`, so I would choose that.

Immediate advantage: PHP code in tests can generate exactly these
JSON samples, rather than dumping slightly different strings for
`date-time` fields.

The diff is just a sample, but it would be easy enough to complete it
with `s/00Z/00+00:00/g` on `src/samples/`.